# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: AI Centralized Minutes API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /auth/login:
    post:
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Logged in }
...